net modular: rewrite to allow for many underlying transports
