+++ kinode/src/eth/default_providers_mainnet.json
@@ -31,8 +31,12 @@
                     "owner": "",
                     "node": "",
                     "public_key": "0x54f5a8a4c625d5925e63ed3f0203b63e007e3f822d7858bd98b1fd9704c99451",
-                    "ip": "147.135.114.167",
-                    "port": 9999,
+                    "ips": [
+                        "147.135.114.167"
+                    ],
+                    "ports": {
+                        "ws": 9999
+                    },
                     "routers": []
                 }
             }
@@ -49,8 +53,12 @@
                     "owner": "",
                     "node": "",
                     "public_key": "0x54f5a8a4c625d5925e63ed3f0203b63e007e3f822d7858bd98b1fd9704c99451",
-                    "ip": "147.135.114.167",
-                    "port": 9999,
+                    "ips": [
+                        "147.135.114.167"
+                    ],
+                    "ports": {
+                        "ws": 9999
+                    },
                     "routers": []
                 }
             }
@@ -67,8 +75,12 @@
                     "owner": "",
                     "node": "",
                     "public_key": "0x54f5a8a4c625d5925e63ed3f0203b63e007e3f822d7858bd98b1fd9704c99451",
-                    "ip": "147.135.114.167",
-                    "port": 9999,
+                    "ips": [
+                        "147.135.114.167"
+                    ],
+                    "ports": {
+                        "ws": 9999
+                    },
                     "routers": []
                 }
             }
